♻️ refactor(transports): relocate validate_api function
- move validate_api function from config.py to transports.py for better modularity and code organization
- update imports in config.py to reflect new location of validate_api function
