+++ src/argoproxy/config.py
@@ -10,6 +10,7 @@ import yaml  # type: ignore
 from loguru import logger
 
 from .utils.misc import get_random_port, is_port_available, make_bar
+from .utils.transports import validate_api
 
 PATHS_TO_TRY = [
     "./config.yaml",
@@ -168,33 +169,6 @@ class ArgoConfig:
         logger.info(make_bar())
         logger.info(self)  # Use the __str__ method for formatted output
         logger.info(make_bar())
-
-
-def validate_api(url: str, username: str, payload: dict, timeout: int = 2) -> bool:
-    """
-    Helper to validate API endpoint connectivity.
-    Args:
-        url (str): The API URL to validate.
-        username (str): The username included in the request payload.
-        payload (dict): The request payload in dictionary format.
-
-    Returns:
-        bool: True if validation succeeds, False otherwise.
-    Raises:
-        ValueError: If validation fails
-    """
-    payload["user"] = username
-    request_data = json.dumps(payload).encode("utf-8")
-    req = urllib.request.Request(
-        url, data=request_data, headers={"Content-Type": "application/json"}
-    )
-    try:
-        with urllib.request.urlopen(req, timeout=timeout) as response:
-            if response.getcode() != 200:
-                raise ValueError(f"API returned status code {response.getcode()}")
-            return True
-    except Exception as e:
-        raise ValueError(f"API validation failed for {url}: {str(e)}") from e
 
 
 def _get_user_port_choice(prompt: str, default_port: int) -> int:
